add main component
This componente is responsible for capturing the repository name typed by the user and make a request to github api. The repository name is appended to a list that is saved to local storage on the browser.

Axios was used for api requests, react-router-dom to manipulate the routes on the application and styled-componenents with react-icons for the visual part.